ingress/requirements.txt
setuptools==72.1.0
praw==7.7.1
typer==0.12.3
loguru==0.7.2
python-dotenv==1.0.1